Change plural Representatives to singular Representative
The area of members is pulled from the breadcrumbs ([-2]) of each
member page. However, whilst the deputies are navigated to
by "Guernsey Deputies > _<Area>_ > _<Member>_" the Alderney
representatives are reached by following "Alderney Representatives >
_<Member>_". Therefore, we end up capturing the plural "Representatives" as
we're taking the second from last bread crumb.

This commit uses a hammer to crack the nut by replacing any
occurrence of the plural with the singular.