Avoid writing empty rows to database
Twice, the re-builder has failed to build Guernsey. On investigation
it has been discovered that on each occasion the database has
contained a row of empty fields.

This commit amends the scraper so that it avoids writing to the
DB when a :name field has not been recorded.

(For reference, the Rollbar error is here:
https://rollbar.com/everypolitician/rebuilder/items/896/)